app/requirements.txt
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
autopep8==1.5.1
boto3==1.12.39
botocore==1.15.39
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-admin-sortable2==0.7.6
django-instagram==0.3.1
django-sortedm2m==3.0.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
Pillow==7.1.1
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
urllib3==1.25.8
wcwidth==0.1.9
whitenoise==5.0.1
wrapt==1.11.2
